// aqi-estimator.client/src/app/shared/interfaces/services/truncate-lengths.interface.ts
import { TruncateLength } from "./truncate-length.interface";

export interface TruncateLengths {
    threshold_xs: TruncateLength;
    threshold_sm: TruncateLength;
    threshold_md: TruncateLength;
    threshold_lg: TruncateLength;
    threshold_xl: TruncateLength;
    threshold_2xl: TruncateLength;
    threshold_3xl: TruncateLength;
    threshold_4xl: TruncateLength;
}